requirements.txt
anyio==3.7.1
asgiref==3.7.2
asyncpg==0.25.0
cffi==1.15.1
click==8.1.6
cryptography==41.0.3
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.75.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.12
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.36
starlette==0.17.1
typing_extensions==4.7.1
uvicorn==0.17.6
